Revert daddr_t to 32 bits while we research the reported problems.
git-svn-id: svn+ssh://svn.freebsd.org/base/head@96679 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/sys/types.h
@@ -126,8 +126,8 @@ typedef	quad_t *	qaddr_t;
 typedef	char *		caddr_t;	/* core address */
 typedef	__const char *	c_caddr_t;	/* core address, pointer to const */
 typedef	__volatile char *v_caddr_t;	/* core address, pointer to volatile */
-typedef	int64_t		daddr_t;	/* disk address */
-typedef	u_int64_t	u_daddr_t;	/* unsigned disk address */
+typedef	int32_t		daddr_t;	/* disk address */
+typedef	u_int32_t	u_daddr_t;	/* unsigned disk address */
 typedef	u_int32_t	fixpt_t;	/* fixed point number */
 
 #ifdef _BSD_GID_T_
